requirements.txt
git+https://github.com/tencent-ailab/IP-Adapter.git
accelerate==1.9.0
black==25.1.0
certifi==2025.7.14
charset-normalizer==3.4.2
click==8.2.1
controlnet_aux==0.0.10
diffusers==0.34.0
filelock==3.18.0
fsspec==2025.7.0
hf-xet==1.1.5
huggingface-hub==0.34.1
idna==3.10
importlib_metadata==8.7.0
invisible-watermark==0.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
mpmath==1.3.0
mypy_extensions==1.1.0
networkx==3.5
numpy==2.2.6
opencv-python==4.12.0.88
packaging==25.0
pathspec==0.12.1
pillow==11.3.0
platformdirs==4.3.8
psutil==7.0.0
PyWavelets==1.8.0
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.4
safetensors==0.5.3
setuptools==80.9.0
sympy==1.14.0
tokenizers==0.21.2
torch==2.7.1
torchvision==0.22.1
tqdm==4.67.1
transformers==4.54.0
typing_extensions==4.14.1
urllib3==2.5.0
zipp==3.23.0
python-dotenv==1.0.0
protobuf==6.*
sentencepiece==0.2.0